% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Function to provide a permutation}
\usage{
shuffle(cards)
}
\arguments{
\item{cards}{a vector}
}
\value{
a vector of reordered cards
}
\description{
This function will provide a permutation of the provided cards.
}
\examples{
shuffle(1:52)

}
